+++ NEWS.md
@@ -112,4 +112,4 @@ ckanr 0.1.0
 
 ### NEW FEATURES
 
-* Releasd to CRAN.
+* Released to CRAN.