+++ NEWS.md
@@ -1,3 +1,5 @@
+# ckanr (development version)
+
 ckanr 0.7.0
 ===========